Split ProgressReporter support singleton
